Clean up SlackTeam.scala
