+++ kifi-backend/modules/common/app/com/keepit/slack/models/SlackTeam.scala
@@ -1,14 +1,9 @@
 package com.keepit.slack.models
 
-import com.keepit.common.cache.{ JsonCacheImpl, FortyTwoCachePlugin, CacheStatistics, Key }
-import com.keepit.common.db.Id
-import com.keepit.common.logging.AccessLog
-import com.keepit.model.Organization
+import com.keepit.common.strings.ValidInt
 import com.kifi.macros.json
-import play.api.libs.json.Json
+import play.api.libs.json._
 import play.api.mvc.{ QueryStringBindable, PathBindable }
-
-import scala.concurrent.duration.Duration
 
 @json case class SlackTeamId(value: String)
 object SlackTeamId {
@@ -32,7 +27,6 @@ object SlackTeamId {
 @json case class SlackTeamName(value: String)
 @json case class SlackTeamDomain(value: String)
 @json case class SlackTeamEmailDomain(value: String)
-@json case class InternalSlackTeamInfo(orgId: Option[Id[Organization]], teamName: SlackTeamName) // augment as needed
 
 
 object SlackTeamIconReads extends Reads[Map[Int, String]] {
@@ -42,11 +36,3 @@ object SlackTeamIconReads extends Reads[Map[Int, String]] {
     if (isDefaultImage) Map.empty[Int, String] else obj.value.collect { case (sizePattern(ValidInt(size)), JsString(imageUrl)) => size -> imageUrl }.toMap
   }
 }
-
-case class InternalSlackTeamInfoKey(id: SlackTeamId) extends Key[InternalSlackTeamInfo] {
-  override val version = 1
-  val namespace = "internal_slack_team_info_by_slack_team_id"
-  def toKey(): String = id.value
-}
-class InternalSlackTeamInfoCache(stats: CacheStatistics, accessLog: AccessLog, innermostPluginSettings: (FortyTwoCachePlugin, Duration), innerToOuterPluginSettings: (FortyTwoCachePlugin, Duration)*)
-  extends JsonCacheImpl[InternalSlackTeamInfoKey, InternalSlackTeamInfo](stats, accessLog, innermostPluginSettings, innerToOuterPluginSettings: _*)(Json.format[InternalSlackTeamInfo])
